+++ packages/core/src/helper/index.ts
@@ -6,3 +6,4 @@ export * from './tokenizer';
 export * from './eventGroups';
 export { selectGridLayout, GRID_LAYOUTS } from './grid-layouts';
 export { setDifference } from './set-helper';
+export { supportsScreenSharing } from './featureDetection';